update | delete with file

+++ src/controllers/manufacturer.js
@@ -45,8 +45,10 @@ exports.create = async (req, res, next) => {
 exports.update = async (req, res, next) => {
   const { body: { id, ...body }, file } = req;
   try {
-    const manufacturer = await Manufacturer.findById(id).exec();
-    if (manufacturer.imagePath !== '/assets/uploads/unnamed.jpg') {
+    const manufacturer = await Manufacturer.findById(id)
+      .exec();
+    if (manufacturer.imagePath !== '/assets/uploads/unnamed.jpg' &&
+      fs.existsSync(manufacturer.imagePath.replace(/assets/, 'src'))) {
       await fs.unlinkSync(`./${manufacturer.imagePath.replace(/assets/, 'src')}`);
     }
     manufacturer.name = body.name || manufacturer.name;
@@ -67,7 +69,8 @@ exports.delete = async (req, res, next) => {
     const manufactures = await Manufacturer.find({ _id: { $in: id } });
     for (let i = 0, length = manufactures.length; i < length; i++) {
       if (manufactures[i]) {
-        if (manufactures[i].imagePath !== '/assets/uploads/unnamed.jpg') {
+        if (manufactures[i].imagePath !== '/assets/uploads/unnamed.jpg' &&
+          fs.existsSync(manufactures[i].imagePath.replace(/assets/, 'src'))) {
           await fs.unlinkSync(`.${manufactures[i].imagePath.replace(/assets/, 'src')}`);
         }
       }
